# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LaptopCondition = typeof LaptopCondition[keyof typeof LaptopCondition];


export const LaptopCondition = {
  new: 'new',
  refurbished: 'refurbished',
} as const;

export interface Laptop {
  id: number;
  /** @nullable */
  sku?: string | null;
  brand: string;
  model: string;
  title: string;
  /** @nullable */
  processor?: string | null;
  /** @nullable */
  processorSpeedMin?: string | null;
  /** @nullable */
  processorSpeedMax?: string | null;
  /** @nullable */
  cores?: number | null;
  /** @nullable */
  threads?: number | null;
  /** @nullable */
  ram?: number | null;
  /** @nullable */
  ramType?: string | null;
  /** @nullable */
  storage?: number | null;
  /** @nullable */
  storageType?: string | null;
  /** @nullable */
  gpu?: string | null;
  /** @nullable */
  operatingSystem?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  screenResolution?: string | null;
  /** @nullable */
  touchscreen?: boolean | null;
  /** @nullable */
  batteryEstimation?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  conditionScore?: number | null;
  /** @nullable */
  conditionDescription?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  price: number;
  stockQuantity: number;
  condition: LaptopCondition;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt: string;
}

export type LaptopInputCondition = typeof LaptopInputCondition[keyof typeof LaptopInputCondition];


export const LaptopInputCondition = {
  new: 'new',
  refurbished: 'refurbished',
} as const;

export interface LaptopInput {
  sku?: string;
  brand: string;
  model: string;
  title: string;
  processor?: string;
  processorSpeedMin?: string;
  processorSpeedMax?: string;
  cores?: number;
  threads?: number;
  ram?: number;
  ramType?: string;
  storage?: number;
  storageType?: string;
  gpu?: string;
  operatingSystem?: string;
  screenSize?: string;
  screenResolution?: string;
  touchscreen?: boolean;
  batteryEstimation?: string;
  weight?: string;
  conditionScore?: number;
  conditionDescription?: string;
  warrantyMonths?: number;
  price: number;
  stockQuantity: number;
  condition: LaptopInputCondition;
  description?: string;
  imageUrl?: string;
  featured?: boolean;
}

export type LaptopUpdateCondition = typeof LaptopUpdateCondition[keyof typeof LaptopUpdateCondition];


export const LaptopUpdateCondition = {
  new: 'new',
  refurbished: 'refurbished',
} as const;

export interface LaptopUpdate {
  brand?: string;
  model?: string;
  title?: string;
  processor?: string;
  price?: number;
  stockQuantity?: number;
  condition?: LaptopUpdateCondition;
  description?: string;
  imageUrl?: string;
  featured?: boolean;
  conditionScore?: number;
  warrantyMonths?: number;
}

export interface LaptopStats {
  totalLaptops: number;
  inStock: number;
  newCount: number;
  refurbishedCount: number;
  brands: number;
}

export interface Accessory {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface AccessoryInput {
  name: string;
  category: string;
  description?: string;
  price: number;
  stock: number;
  imageUrl?: string;
}

export interface BrandSummary {
  brand: string;
  count: number;
  imageUrl: string;
  tag: string;
}

export type ListLaptopsParams = {
brand?: string;
condition?: ListLaptopsCondition;
minPrice?: number;
maxPrice?: number;
ram?: number;
search?: string;
inStock?: boolean;
};

export type ListLaptopsCondition = typeof ListLaptopsCondition[keyof typeof ListLaptopsCondition];


export const ListLaptopsCondition = {
  new: 'new',
  refurbished: 'refurbished',
} as const;

export type ListAccessoriesParams = {
category?: string;
};