# lib/api-zod/src/generated/types/accessory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Accessory {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
}