# lib/api-zod/src/generated/types/accessoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessoryInput {
  name: string;
  category: string;
  description?: string;
  price: number;
  stock: number;
  imageUrl?: string;
}