# lib/api-zod/src/generated/types/brandSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandSummary {
  brand: string;
  count: number;
  imageUrl: string;
  tag: string;
}