# lib/api-zod/src/generated/types/laptop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */
import type { LaptopCondition } from './laptopCondition';

export interface Laptop {
  id: number;
  /** @nullable */
  sku?: string | null;
  brand: string;
  model: string;
  title: string;
  /** @nullable */
  processor?: string | null;
  /** @nullable */
  processorSpeedMin?: string | null;
  /** @nullable */
  processorSpeedMax?: string | null;
  /** @nullable */
  cores?: number | null;
  /** @nullable */
  threads?: number | null;
  /** @nullable */
  ram?: number | null;
  /** @nullable */
  ramType?: string | null;
  /** @nullable */
  storage?: number | null;
  /** @nullable */
  storageType?: string | null;
  /** @nullable */
  gpu?: string | null;
  /** @nullable */
  operatingSystem?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  screenResolution?: string | null;
  /** @nullable */
  touchscreen?: boolean | null;
  /** @nullable */
  batteryEstimation?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  conditionScore?: number | null;
  /** @nullable */
  conditionDescription?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  price: number;
  stockQuantity: number;
  condition: LaptopCondition;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt: string;
}