# lib/api-zod/src/generated/types/laptopInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */
import type { LaptopInputCondition } from './laptopInputCondition';

export interface LaptopInput {
  sku?: string;
  brand: string;
  model: string;
  title: string;
  processor?: string;
  processorSpeedMin?: string;
  processorSpeedMax?: string;
  cores?: number;
  threads?: number;
  ram?: number;
  ramType?: string;
  storage?: number;
  storageType?: string;
  gpu?: string;
  operatingSystem?: string;
  screenSize?: string;
  screenResolution?: string;
  touchscreen?: boolean;
  batteryEstimation?: string;
  weight?: string;
  conditionScore?: number;
  conditionDescription?: string;
  warrantyMonths?: number;
  price: number;
  stockQuantity: number;
  condition: LaptopInputCondition;
  description?: string;
  imageUrl?: string;
  featured?: boolean;
}