# lib/api-zod/src/generated/types/laptopStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */

export interface LaptopStats {
  totalLaptops: number;
  inStock: number;
  newCount: number;
  refurbishedCount: number;
  brands: number;
}