# lib/api-zod/src/generated/types/laptopUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROC DZ Laptop Store API
 * OpenAPI spec version: 0.1.0
 */
import type { LaptopUpdateCondition } from './laptopUpdateCondition';

export interface LaptopUpdate {
  brand?: string;
  model?: string;
  title?: string;
  processor?: string;
  price?: number;
  stockQuantity?: number;
  condition?: LaptopUpdateCondition;
  description?: string;
  imageUrl?: string;
  featured?: boolean;
  conditionScore?: number;
  warrantyMonths?: number;
}